backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
langchain
langchain-community
fastembed
openai
